third_party/WebKit/LayoutTests/virtual/pointerevent/fast/events/pointerevents/mouse-pointer-event-properties-expected.txt

Verifies that pointer event parameters are correct when fired for mouse events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Received pointerover
Received mouseover
PASS lastPointerEvent is non-null.
PASS lastPointerEvent.type is "pointerover"
PASS lastPointerEvent.bubbles is true
PASS lastPointerEvent.cancelable is true
PASS lastPointerEvent.pointerId is 0
PASS lastPointerEvent.pointerType is "mouse"
PASS lastPointerEvent.isPrimary is true
PASS lastPointerEvent.clientX is 23
PASS lastPointerEvent.clientY is 31
PASS lastPointerEvent.layerX is 23
PASS lastPointerEvent.layerY is 31
PASS lastPointerEvent.movementX is 0
PASS lastPointerEvent.movementY is 0
PASS lastPointerEvent.offsetX is 5
PASS lastPointerEvent.offsetY is 5
PASS lastPointerEvent.pageX is 23
PASS lastPointerEvent.pageY is 31
PASS lastPointerEvent.screenX is 23
PASS lastPointerEvent.screenY is 31
PASS lastPointerEvent.x is 23
PASS lastPointerEvent.y is 31
PASS lastPointerEvent.button is 0
PASS lastPointerEvent.buttons is 0
PASS lastPointerEvent.getModifierState('Control'); is false
PASS lastPointerEvent.getModifierState('Shift'); is false
PASS lastPointerEvent.getModifierState('Meta'); is false
PASS lastPointerEvent.getModifierState('Alt'); is false
PASS lastPointerEvent.getModifierState('AltGraph'); is false
PASS lastPointerEvent.getModifierState('OS'); is false
PASS lastPointerEvent.getModifierState('Fn'); is false
PASS lastPointerEvent.getModifierState('CapsLock'); is false
PASS lastPointerEvent.getModifierState('ScrollLock'); is false
PASS lastPointerEvent.getModifierState('NumLock'); is false
PASS lastPointerEvent.getModifierState('Symbol'); is false
Received pointerenter
Received mouseenter
PASS lastPointerEvent is non-null.
PASS lastPointerEvent.type is "pointerenter"
PASS lastPointerEvent.bubbles is false
PASS lastPointerEvent.cancelable is false
PASS lastPointerEvent.pointerId is 0
PASS lastPointerEvent.pointerType is "mouse"
PASS lastPointerEvent.isPrimary is true
PASS lastPointerEvent.clientX is 23
PASS lastPointerEvent.clientY is 31
PASS lastPointerEvent.layerX is 23
PASS lastPointerEvent.layerY is 31
PASS lastPointerEvent.movementX is 0
PASS lastPointerEvent.movementY is 0
PASS lastPointerEvent.offsetX is 5
PASS lastPointerEvent.offsetY is 5
PASS lastPointerEvent.pageX is 23
PASS lastPointerEvent.pageY is 31
PASS lastPointerEvent.screenX is 23
PASS lastPointerEvent.screenY is 31
PASS lastPointerEvent.x is 23
PASS lastPointerEvent.y is 31
PASS lastPointerEvent.button is 0
PASS lastPointerEvent.buttons is 0
PASS lastPointerEvent.getModifierState('Control'); is false
PASS lastPointerEvent.getModifierState('Shift'); is false
PASS lastPointerEvent.getModifierState('Meta'); is false
PASS lastPointerEvent.getModifierState('Alt'); is false
PASS lastPointerEvent.getModifierState('AltGraph'); is false
PASS lastPointerEvent.getModifierState('OS'); is false
PASS lastPointerEvent.getModifierState('Fn'); is false
PASS lastPointerEvent.getModifierState('CapsLock'); is false
PASS lastPointerEvent.getModifierState('ScrollLock'); is false
PASS lastPointerEvent.getModifierState('NumLock'); is false
PASS lastPointerEvent.getModifierState('Symbol'); is false
Received pointerout
Received mouseout
PASS lastPointerEvent is non-null.
PASS lastPointerEvent.type is "pointerout"
PASS lastPointerEvent.bubbles is true
PASS lastPointerEvent.cancelable is true
PASS lastPointerEvent.pointerId is 0
PASS lastPointerEvent.pointerType is "mouse"
PASS lastPointerEvent.isPrimary is true
PASS lastPointerEvent.clientX is 13
PASS lastPointerEvent.clientY is 21
PASS lastPointerEvent.layerX is 13
PASS lastPointerEvent.layerY is 21
PASS lastPointerEvent.movementX is 0
PASS lastPointerEvent.movementY is 0
PASS lastPointerEvent.offsetX is -5
PASS lastPointerEvent.offsetY is -5
PASS lastPointerEvent.pageX is 13
PASS lastPointerEvent.pageY is 21
PASS lastPointerEvent.screenX is 13
PASS lastPointerEvent.screenY is 21
PASS lastPointerEvent.x is 13
PASS lastPointerEvent.y is 21
PASS lastPointerEvent.button is 0
PASS lastPointerEvent.buttons is 0
PASS lastPointerEvent.getModifierState('Control'); is false
PASS lastPointerEvent.getModifierState('Shift'); is false
PASS lastPointerEvent.getModifierState('Meta'); is false
PASS lastPointerEvent.getModifierState('Alt'); is false
PASS lastPointerEvent.getModifierState('AltGraph'); is false
PASS lastPointerEvent.getModifierState('OS'); is false
PASS lastPointerEvent.getModifierState('Fn'); is false
PASS lastPointerEvent.getModifierState('CapsLock'); is false
PASS lastPointerEvent.getModifierState('ScrollLock'); is false
PASS lastPointerEvent.getModifierState('NumLock'); is false
PASS lastPointerEvent.getModifierState('Symbol'); is false
Received pointerleave
Received mouseleave
PASS lastPointerEvent is non-null.
PASS lastPointerEvent.type is "pointerleave"
PASS lastPointerEvent.bubbles is false
PASS lastPointerEvent.cancelable is false
PASS lastPointerEvent.pointerId is 0
PASS lastPointerEvent.pointerType is "mouse"
PASS lastPointerEvent.isPrimary is true
PASS lastPointerEvent.clientX is 13
PASS lastPointerEvent.clientY is 21
PASS lastPointerEvent.layerX is 13
PASS lastPointerEvent.layerY is 21
PASS lastPointerEvent.movementX is 0
PASS lastPointerEvent.movementY is 0
PASS lastPointerEvent.offsetX is -5
PASS lastPointerEvent.offsetY is -5
PASS lastPointerEvent.pageX is 13
PASS lastPointerEvent.pageY is 21
PASS lastPointerEvent.screenX is 13
PASS lastPointerEvent.screenY is 21
PASS lastPointerEvent.x is 13
PASS lastPointerEvent.y is 21
PASS lastPointerEvent.button is 0
PASS lastPointerEvent.buttons is 0
PASS lastPointerEvent.getModifierState('Control'); is false
PASS lastPointerEvent.getModifierState('Shift'); is false
PASS lastPointerEvent.getModifierState('Meta'); is false
PASS lastPointerEvent.getModifierState('Alt'); is false
PASS lastPointerEvent.getModifierState('AltGraph'); is false
PASS lastPointerEvent.getModifierState('OS'); is false
PASS lastPointerEvent.getModifierState('Fn'); is false
PASS lastPointerEvent.getModifierState('CapsLock'); is false
PASS lastPointerEvent.getModifierState('ScrollLock'); is false
PASS lastPointerEvent.getModifierState('NumLock'); is false
PASS lastPointerEvent.getModifierState('Symbol'); is false
PASS successfullyParsed is true

TEST COMPLETE
